fabric/requirements.txt
amqp==5.1.1
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
blessed==1.20.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.11.2
django-picklefield==3.1
django-q==1.3.9
django-redis==5.4.0
django-rest-swagger==2.2.0
django-swagger-ui==0.1.11
django-templated-mail==1.1.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jinxed==1.2.0
kombu==5.3.2
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8

